tests/integration/requirements.txt
pytest>=7.0.0
requests>=2.28.0
pyyaml>=6.0
pytest-timeout>=2.1.0